requirements.txt
aioredis==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.0
charset-normalizer==3.0.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.9
django-cors-headers==3.13.0
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.0.0
hiredis==2.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.4
oauthlib==3.2.2
Pillow==9.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
zope.interface==5.5.2
